refactor(ledger): move gas accounting into module
